Refine advanced prompt completion
Adjust the behavior of the autocompletion feature in the advanced TUI
prompt for UX reasons.

Removing `WithShowCompletionAtStart` aligns the behavior with what happens
when the user executes a command.
Additionally, the maximum number of suggestions is reduced from 100
to 10. A list of 100 items is difficult to navigate and clutters the
interface.
Both of these changes take into account the newly-added support for paging
through suggestions.